Fix BottomNavigationBar type error
- Resolved argument_type_not_assignable error for fontSize parameters
- Added explicit toDouble() calls to handle nullable double type
- Bottom navigation component now fully functional

+++ flutter/glue_flutter/lib/src/lib/ui/material/widgets/bottom_navigation_bar.dart
@@ -33,8 +33,10 @@ Eval<Ir> _createBottomNavigationBar(WidgetProperties properties) {
     unselectedIconTheme: properties.getValue('unselected-icon-theme'),
     selectedLabelStyle: properties.getValue('selected-label-style'),
     unselectedLabelStyle: properties.getValue('unselected-label-style'),
-    selectedFontSize: properties.getDouble('selected-font-size') ?? 14.0,
-    unselectedFontSize: properties.getDouble('unselected-font-size') ?? 12.0,
+    selectedFontSize: (properties.getDouble('selected-font-size') ?? 14.0)
+        .toDouble(),
+    unselectedFontSize: (properties.getDouble('unselected-font-size') ?? 12.0)
+        .toDouble(),
     showSelectedLabels: properties.getBool('show-selected-labels'),
     showUnselectedLabels: properties.getBool('show-unselected-labels'),
     enableFeedback: properties.getBool('enable-feedback'),
